Sky box and IAP packeg added

+++ Assets/Mobile Farmer/Scripts/Utility/JY_Mono_Utiliy.cs
@@ -156,7 +156,7 @@ public class BarnItem{
 public struct S_PurchessAmount_Pair
 {
     public int amountToCredit;
-    public int amountToDebit;
+    public float amountToDebit;
 }
 [System.Serializable]
 public struct WorkerContiner
